// src/client/model/accessTokenDto.ts
/**
 * PlanQK Platform OpenAPI definition
 * OpenAPI definition for the Platform and Ecosystem for Quantum-assisted Artificial Intelligence Platform
 *
 * The version of the OpenAPI document: v1
 * Contact: info@stoneone.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AccessTokenDto {
    'id'?: string;
    'name'?: string;
    'scopes'?: Set<AccessTokenDto.ScopesEnum>;
    'createdAt'?: Date;
    'usedAt'?: Date;
    'expiresAt'?: string;
    'value'?: string;
    'token'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "scopes",
            "baseName": "scopes",
            "type": "Set<AccessTokenDto.ScopesEnum>"
        },
        {
            "name": "createdAt",
            "baseName": "createdAt",
            "type": "Date"
        },
        {
            "name": "usedAt",
            "baseName": "usedAt",
            "type": "Date"
        },
        {
            "name": "expiresAt",
            "baseName": "expiresAt",
            "type": "string"
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "string"
        },
        {
            "name": "token",
            "baseName": "token",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AccessTokenDto.attributeTypeMap;
    }
}

export namespace AccessTokenDto {
    export enum ScopesEnum {
        Api = <any> 'API',
        QuantumTokens = <any> 'QUANTUM_TOKENS'
    }
}