// src/client/model/createInternalSubscriptionRequest.ts
/**
 * PlanQK Platform OpenAPI definition
 * OpenAPI definition for the Platform and Ecosystem for Quantum-assisted Artificial Intelligence Platform
 *
 * The version of the OpenAPI document: v1
 * Contact: info@stoneone.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateInternalSubscriptionRequest {
    'applicationId'?: string;
    'serviceId'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "applicationId",
            "baseName": "applicationId",
            "type": "string"
        },
        {
            "name": "serviceId",
            "baseName": "serviceId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateInternalSubscriptionRequest.attributeTypeMap;
    }
}