// src/client/model/createManagedServicePricingPlanRequest.ts
/**
 * PlanQK Platform OpenAPI definition
 * OpenAPI definition for the Platform and Ecosystem for Quantum-assisted Artificial Intelligence Platform
 *
 * The version of the OpenAPI document: v1
 * Contact: info@stoneone.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export class CreateManagedServicePricingPlanRequest {
    'type': CreateManagedServicePricingPlanRequest.TypeEnum;
    'profitPercentage'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "type",
            "baseName": "type",
            "type": "CreateManagedServicePricingPlanRequest.TypeEnum"
        },
        {
            "name": "profitPercentage",
            "baseName": "profitPercentage",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CreateManagedServicePricingPlanRequest.attributeTypeMap;
    }
}

export namespace CreateManagedServicePricingPlanRequest {
    export enum TypeEnum {
        Free = <any> 'FREE',
        Commercial = <any> 'COMMERCIAL'
    }
}