// src/client/model/createProductRequest.ts
/**
 * PlanQK Platform OpenAPI definition
 * OpenAPI definition for the Platform and Ecosystem for Quantum-assisted Artificial Intelligence Platform
 *
 * The version of the OpenAPI document: v1
 * Contact: info@stoneone.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export class CreateProductRequest {
    'name': string;
    'unitLabel'?: string;
    'amountMajor': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "unitLabel",
            "baseName": "unitLabel",
            "type": "string"
        },
        {
            "name": "amountMajor",
            "baseName": "amountMajor",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CreateProductRequest.attributeTypeMap;
    }
}